requirements.txt
langchain
openai
faiss-cpu
chromadb
python-dotenv
pypdf
duckduckgo-search
streamlit
watchdog